refactor routes and models into modules
