+++ rockr/auth.py
@@ -1,6 +1,6 @@
-
 def login(data):
     return "Login"
+
 
 def logout(data):
     return "Logout"
