+++ rockr/views.py
@@ -1,0 +1,48 @@
+from flask import request
+from rockr import auth, app
+from rockr.db.models import bands, users
+
+
+def format_response(status, data):
+    return {"status": status, "data": data}
+
+
+@app.route('/', methods=["GET"])
+def index():
+    return "Welcome to Rockr!"
+
+
+@app.route('/login', methods=["POST"])
+def login():
+    data = auth.login(request.json)
+    return format_response(200, data)
+
+
+@app.route('/logout')
+def logout():
+    data = auth.logout(request.json)
+    return format_response(200, data)
+
+
+@app.route('/register', methods=['POST'])
+def register():
+    data = users.create_user(request.json)
+    return format_response(200, data)
+
+
+@app.route('/get_users', methods=["GET"])
+def get_users():
+    data = users.get_users()
+    return format_response(200, data)
+
+
+@app.route('/get_bands', methods=["GET"])
+def get_bands():
+    data = bands.get_bands()
+    return format_response(200, data)
+
+
+@app.route('/get_user_role', methods=["POST"])
+def get_user_role():
+    data = users.get_user_role(request.json)
+    return format_response(200, data)
